requirements.txt
colorama==0.4.3
Flask==2.2.3
Flask_Cors==3.0.10
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==9.4.0
pyfirmata==1.1.0
pytesseract==0.3.10
requests==2.22.0